requirements.txt
pytest-django
Django
django-rest-swagger
djangorestframework
drf-spectacular
djangorestframework-simplejwt